Add a SHA param
This allows specifying which commit the status check is for. It
preserves the current default (the HEAD SHA in pull requests) but
falls back to `github.sha` which is more generally available (on
`push` events, for example).

This also allows overriding the default completly for other use
cases (e.g. a scheduled task updating commit statuses).